case_study/requirements.txt
flask==3.0.0
cassandra-driver==3.29.0
passlib==1.7.4